Move external and bplan to mB dashboard
