+++ meinberlin/apps/dashboard2/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ObjectDoesNotExist
-from django.core.exceptions import ValidationError
 from django.forms import inlineformset_factory
 from django.utils.translation import ugettext_lazy as _
 
@@ -10,9 +9,7 @@ from adhocracy4.maps import models as map_models
 from adhocracy4.modules import models as module_models
 from adhocracy4.phases import models as phase_models
 from adhocracy4.projects import models as project_models
-from meinberlin.apps.bplan import models as bplan_models
 from meinberlin.apps.datetimefield import fields as datetime_fields
-from meinberlin.apps.extprojects import models as extproject_models
 from meinberlin.apps.maps.widgets import MapChoosePolygonWithPresetWidget
 from meinberlin.apps.users.fields import CommaSeparatedEmailField
 
@@ -201,73 +198,3 @@ CategoryFormSet = inlineformset_factory(module_models.Module,
                                         form=CategoryForm,
                                         formset=ModuleDashboardFormSet,
                                         )
-
-
-# meinBerlin related Forms
-
-class ExternalProjectCreateForm(ProjectCreateForm):
-
-    class Meta:
-        model = extproject_models.ExternalProject
-        fields = ['name', 'description', 'tile_image', ]
-
-
-class ExternalProjectForm(ProjectDashboardForm):
-
-    start_date = datetime_fields.DateTimeField(
-        time_format='%H:%M',
-        required=False,
-        require_all_fields=False,
-        label=(_('Start date'), _('Start time'))
-    )
-    end_date = datetime_fields.DateTimeField(
-        time_format='%H:%M',
-        required=False,
-        require_all_fields=False,
-        label=(_('End date'), _('End time'))
-    )
-
-    class Meta:
-        model = extproject_models.ExternalProject
-        fields = ['name', 'url', 'description', 'tile_image', 'is_archived']
-        required_for_project_publish = ['name', 'url', 'description']
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        self.initial['start_date'] = self.instance.phase.start_date
-        self.initial['end_date'] = self.instance.phase.end_date
-
-    def clean_end_date(self, *args, **kwargs):
-        start_date = self.cleaned_data['start_date']
-        end_date = self.cleaned_data['end_date']
-        if start_date and end_date and end_date < start_date:
-            raise ValidationError(
-                _('End date can not be smaller than the start date.'))
-        return end_date
-
-    def save(self, commit=True):
-        project = super().save(commit)
-
-        if commit:
-            phase = project.phase
-            phase.start_date = self.cleaned_data['start_date']
-            phase.end_date = self.cleaned_data['end_date']
-            phase.save()
-
-
-class BplanProjectCreateForm(ExternalProjectCreateForm):
-
-    class Meta:
-        model = bplan_models.Bplan
-        fields = ['name', 'description', 'tile_image', ]
-
-
-class BplanProjectForm(ExternalProjectForm):
-
-    class Meta:
-        model = bplan_models.Bplan
-        fields = ['name', 'url', 'description', 'tile_image', 'is_archived',
-                  'office_worker_email']
-        required_for_project_publish = ['name', 'url', 'description',
-                                        'office_worker_email']
