+++ meinberlin/apps/dashboard2/views.py
@@ -15,8 +15,6 @@ from django.views.generic.detail import SingleObjectMixin
 from adhocracy4.modules import models as module_models
 from adhocracy4.phases import models as phase_models
 from adhocracy4.projects import models as project_models
-from meinberlin.apps.bplan import models as bplan_models
-from meinberlin.apps.extprojects import models as extproject_models
 from meinberlin.apps.projects.emails import InviteParticipantEmail
 
 from . import blueprints
@@ -24,7 +22,6 @@ from . import forms
 from . import get_project_dashboard
 from . import mixins
 from . import signals
-from .components.forms.views import ProjectComponentFormView
 
 User = get_user_model()
 
@@ -296,58 +293,3 @@ class DashboardProjectParticipantsView(AbstractProjectUserListView):
                                         participant_ids=participant_ids)
 
         return response
-
-
-# meinBerlin related Views.
-class ExternalProjectCreateView(ProjectCreateView):
-
-    model = extproject_models.ExternalProject
-    slug_url_kwarg = 'project_slug'
-    blueprint_key = 'external-project'
-    form_class = forms.ExternalProjectCreateForm
-    template_name = 'meinberlin_dashboard2/external_project_create_form.html'
-
-
-class ExternalProjectUpdateView(ProjectComponentFormView):
-
-    model = extproject_models.ExternalProject
-
-    def get_project(self, *args, **kwargs):
-        project = super().get_project(*args, **kwargs)
-        return project.externalproject
-
-    def get_object(self, queryset=None):
-        return self.project
-
-    def validate_object_project(self):
-        return True
-
-    def validate_object_module(self):
-        return True
-
-
-class BplanProjectCreateView(ExternalProjectCreateView):
-
-    model = bplan_models.Bplan
-    slug_url_kwarg = 'project_slug'
-    blueprint_key = 'bplan'
-    form_class = forms.BplanProjectCreateForm
-    template_name = 'meinberlin_dashboard2/external_project_create_form.html'
-
-
-class BplanProjectUpdateView(ProjectComponentFormView):
-
-    model = bplan_models.Bplan
-
-    def get_project(self, *args, **kwargs):
-        project = super().get_project(*args, **kwargs)
-        return project.externalproject.bplan
-
-    def get_object(self, queryset=None):
-        return self.project
-
-    def validate_object_project(self):
-        return True
-
-    def validate_object_module(self):
-        return True
